+++ tests/jax/multi_chip/bounties/falcon3_7b/test/generate_flax.py
@@ -16,7 +16,7 @@ def main(model_name: str, prompt: str):
     # Example usage
     config = AutoConfig.from_pretrained(
         model_name,
-        num_hidden_layers=32, # can be reduced for faster runtime
+        num_hidden_layers=28, # can be reduced for faster runtime
         torch_dtype=torch.float32
     )
